+++ apps/modern-component-data-fetch/host/src/routes/page.tsx
@@ -1,0 +1,9 @@
+const Index = () => {
+  return (
+    <>
+      <h1>Welcome message from host</h1>
+    </>
+  );
+};
+
+export default Index;