+++ packages/modernjs/src/cli/configPlugin.ts
@@ -4,8 +4,12 @@ import { moduleFederationPlugin, encodeName } from '@module-federation/sdk';
 import { bundle } from '@modern-js/node-bundle-require';
 import { PluginOptions } from '../types';
 import { LOCALHOST, PLUGIN_IDENTIFIER } from '../constant';
-import { autoDeleteSplitChunkCacheGroups } from '@module-federation/rsbuild-plugin/utils';
-import logger from './logger';
+import {
+  autoDeleteSplitChunkCacheGroups,
+  addDataFetchExposes,
+} from '@module-federation/rsbuild-plugin/utils';
+import logger from '../logger';
+import { isDev } from './utils';
 
 import type { InternalModernPluginOptions } from '../types';
 import type {
@@ -19,7 +23,6 @@ import type {
 import type { BundlerChainConfig } from '../interfaces/bundler';
 
 const defaultPath = path.resolve(process.cwd(), 'module-federation.config.ts');
-const isDev = process.env.NODE_ENV === 'development';
 
 export type ConfigType<T> = T extends 'webpack'
   ? webpack.Configuration
@@ -144,14 +147,19 @@ export const patchMFConfig = (
   mfConfig: moduleFederationPlugin.ModuleFederationPluginOptions,
   isServer: boolean,
   remoteIpStrategy?: 'ipv4' | 'inherit',
+  enableSSR?: boolean,
 ) => {
   replaceRemoteUrl(mfConfig, remoteIpStrategy);
+  addDataFetchExposes(mfConfig.exposes, isServer);
+
   if (mfConfig.remoteType === undefined) {
     mfConfig.remoteType = 'script';
   }
+
   if (!mfConfig.name) {
     throw new Error(`${PLUGIN_IDENTIFIER} mfConfig.name can not be empty!`);
   }
+
   const runtimePlugins = [...(mfConfig.runtimePlugins || [])];
 
   patchDTSConfig(mfConfig, isServer);
@@ -161,7 +169,12 @@ export const patchMFConfig = (
     runtimePlugins,
   );
 
-  if (isDev) {
+  injectRuntimePlugins(
+    require.resolve('@module-federation/modern-js/auto-fetch-data'),
+    runtimePlugins,
+  );
+
+  if (enableSSR && isDev()) {
     injectRuntimePlugins(
       require.resolve('@module-federation/modern-js/resolve-entry-ipv4'),
       runtimePlugins,
@@ -173,7 +186,7 @@ export const patchMFConfig = (
       require.resolve('@module-federation/node/runtimePlugin'),
       runtimePlugins,
     );
-    if (isDev) {
+    if (isDev()) {
       injectRuntimePlugins(
         require.resolve(
           '@module-federation/node/record-dynamic-remote-entry-hash-plugin',
@@ -324,7 +337,7 @@ export function patchBundlerConfig(options: {
     );
   }
 
-  if (isDev && chain.output.get('publicPath') === 'auto') {
+  if (isDev() && chain.output.get('publicPath') === 'auto') {
     // TODO: only in dev temp
     const port =
       modernjsConfig.dev?.port || modernjsConfig.server?.port || 8080;
@@ -340,12 +353,12 @@ export function patchBundlerConfig(options: {
       uniqueName &&
       !chunkFileName.includes(uniqueName)
     ) {
-      const suffix = `${encodeName(uniqueName)}-[chunkhash].js`;
+      const suffix = `${encodeName(uniqueName)}-[contenthash].js`;
       chain.output.chunkFilename(chunkFileName.replace('.js', suffix));
     }
   }
   // modernjs project has the same entry for server/client, add polyfill:false to skip compile error in browser target
-  if (isDev && enableSSR && !isServer) {
+  if (isDev() && enableSSR && !isServer) {
     chain.resolve.fallback
       .set('crypto', false)
       .set('stream', false)
@@ -385,7 +398,6 @@ export const moduleFederationConfigPlugin = (
         return;
       }
       const isWeb = isWebTarget(target);
-      // @ts-expect-error chain type is not correct
       addMyTypes2Ignored(chain, !isWeb ? ssrConfig : csrConfig);
 
       const targetMFConfig = !isWeb ? ssrConfig : csrConfig;
@@ -393,10 +405,10 @@ export const moduleFederationConfigPlugin = (
         targetMFConfig,
         !isWeb,
         userConfig.remoteIpStrategy || 'ipv4',
+        enableSSR,
       );
 
       patchBundlerConfig({
-        // @ts-expect-error chain type is not correct
         chain,
         isServer: !isWeb,
         modernjsConfig,
@@ -451,6 +463,12 @@ export const moduleFederationConfigPlugin = (
             'Access-Control-Allow-Headers': '*',
           }
         : undefined;
+      const defineConfig = {
+        REMOTE_IP_STRATEGY: JSON.stringify(userConfig.remoteIpStrategy),
+      };
+      if (enableSSR && isDev()) {
+        defineConfig['FEDERATION_IPV4'] = JSON.stringify(ipv4);
+      }
       return {
         tools: {
           devServer: {
@@ -464,10 +482,7 @@ export const moduleFederationConfigPlugin = (
               '@module-federation/modern-js/runtime',
             ),
           },
-          define: {
-            FEDERATION_IPV4: JSON.stringify(ipv4),
-            REMOTE_IP_STRATEGY: JSON.stringify(userConfig.remoteIpStrategy),
-          },
+          define: defineConfig,
           enableAsyncEntry:
             bundlerType === 'rspack'
               ? (modernjsConfig.source?.enableAsyncEntry ?? true)