+++ packages/rsbuild-plugin/src/cli/index.ts
@@ -4,33 +4,64 @@ import {
   PLUGIN_NAME,
 } from '@module-federation/enhanced/rspack';
 import { isRequiredVersion } from '@module-federation/sdk';
-
-import { isRegExp, autoDeleteSplitChunkCacheGroups } from '../utils/index';
 import pkgJson from '../../package.json';
+import logger from '../logger';
+import {
+  isRegExp,
+  autoDeleteSplitChunkCacheGroups,
+  addDataFetchExposes,
+  createSSRMFConfig,
+  createSSRREnvConfig,
+  setSSREnv,
+  SSR_ENV_NAME,
+  SSR_DIR,
+  updateStatsAndManifest,
+  patchSSRRspackConfig,
+} from '../utils';
 
 import type {
   moduleFederationPlugin,
   sharePlugin,
 } from '@module-federation/sdk';
 import type { RsbuildConfig, RsbuildPlugin, Rspack } from '@rsbuild/core';
-import logger from '../logger';
+import {
+  CALL_NAME_MAP,
+  RSPRESS_BUNDLER_CONFIG_NAME,
+  RSPRESS_SSR_DIR,
+} from '../constant';
 
 type ModuleFederationOptions =
   moduleFederationPlugin.ModuleFederationPluginOptions;
 
-export type { ModuleFederationOptions };
+type RSBUILD_PLUGIN_OPTIONS = {
+  ssr?: boolean;
+  // ssr dir, default is ssr
+  ssrDir?: string;
+  // target copy environment name, default is mf
+  environment?: string;
+};
+
+type ExposedAPIType = {
+  options: {
+    nodePlugin?: ModuleFederationPlugin;
+    browserPlugin?: ModuleFederationPlugin;
+    rspressSSGPlugin?: ModuleFederationPlugin;
+    distOutputDir?: string;
+  };
+  isSSRConfig: typeof isSSRConfig;
+  isRspressSSGConfig: typeof isRspressSSGConfig;
+};
+export type { ModuleFederationOptions, ExposedAPIType };
 
 const RSBUILD_PLUGIN_MODULE_FEDERATION_NAME =
   'rsbuild:module-federation-enhanced';
-const RSPACK_PLUGIN_MODULE_FEDERATION_NAME = 'module-federation';
+const RSBUILD_PLUGIN_NAME = '@module-federation/rsbuild-plugin';
 
-export {
-  RSBUILD_PLUGIN_MODULE_FEDERATION_NAME,
-  RSPACK_PLUGIN_MODULE_FEDERATION_NAME,
-  PLUGIN_NAME,
-};
+export { RSBUILD_PLUGIN_MODULE_FEDERATION_NAME, PLUGIN_NAME, SSR_DIR };
 
 const LIB_FORMAT = ['umd', 'modern-module'];
+
+const DEFAULT_MF_ENVIRONMENT_NAME = 'mf';
 
 function isStoryBook(rsbuildConfig: RsbuildConfig) {
   if (
@@ -45,6 +76,9 @@ function isStoryBook(rsbuildConfig: RsbuildConfig) {
 
 export function isMFFormat(bundlerConfig: Rspack.Configuration) {
   const library = bundlerConfig.output?.library;
+  if (bundlerConfig.name === SSR_ENV_NAME) {
+    return true;
+  }
 
   return !(
     typeof library === 'object' &&
@@ -55,11 +89,55 @@ export function isMFFormat(bundlerConfig: Rspack.Configuration) {
   );
 }
 
+const isSSRConfig = (bundlerConfigName?: string) =>
+  Boolean(bundlerConfigName === SSR_ENV_NAME);
+
+const isRspressSSGConfig = (bundlerConfigName?: string) => {
+  return bundlerConfigName === RSPRESS_BUNDLER_CONFIG_NAME;
+};
+
 export const pluginModuleFederation = (
   moduleFederationOptions: ModuleFederationOptions,
+  rsbuildOptions: RSBUILD_PLUGIN_OPTIONS,
 ): RsbuildPlugin => ({
   name: RSBUILD_PLUGIN_MODULE_FEDERATION_NAME,
   setup: (api) => {
+    const {
+      ssr = undefined,
+      ssrDir = SSR_DIR,
+      environment = DEFAULT_MF_ENVIRONMENT_NAME,
+    } = rsbuildOptions || {};
+    const { callerName } = api.context;
+    const originalRsbuildConfig = api.getRsbuildConfig();
+    if (!callerName) {
+      throw new Error(
+        '`callerName` is undefined. Please ensure the @rsbuild/core version is higher than 1.3.21 .',
+      );
+    }
+    const isRslib = callerName === CALL_NAME_MAP.RSLIB;
+    const isRspress = callerName === CALL_NAME_MAP.RSPRESS;
+    const isSSR = Boolean(ssr);
+
+    if (isSSR && !isStoryBook(originalRsbuildConfig)) {
+      if (!isRslib && !isRspress) {
+        throw new Error(`'ssr' option is only supported in rslib.`);
+      }
+      const rsbuildConfig = api.getRsbuildConfig();
+
+      if (
+        !rsbuildConfig.environments?.[environment] ||
+        Object.keys(rsbuildConfig.environments).some(
+          (key) => key.startsWith(environment) && key !== environment,
+        )
+      ) {
+        throw new Error(
+          `Please set ${RSBUILD_PLUGIN_NAME} as global plugin in rslib.config.ts if you set 'ssr:true' .`,
+        );
+      }
+
+      setSSREnv();
+    }
+
     const sharedOptions: [string, sharePlugin.SharedConfig][] = parseOptions(
       moduleFederationOptions.shared || [],
       (item, key) => {
@@ -83,14 +161,103 @@ export const pluginModuleFederation = (
       shared[0].endsWith('/') ? shared[0].slice(0, -1) : shared[0],
     );
 
+    api.modifyRsbuildConfig((config) => {
+      // skip storybook
+      if (isStoryBook(config)) {
+        return;
+      }
+
+      // Change some default configs for remote modules
+      if (moduleFederationOptions.exposes) {
+        config.dev ||= {};
+        config.server ||= {};
+        const userConfig = api.getRsbuildConfig('original');
+
+        // Allow remote modules to be loaded by setting CORS headers
+        // This is required for MF to work properly across different origins
+        config.server.headers ||= {};
+        if (
+          !config.server.headers['Access-Control-Allow-Origin'] &&
+          !(
+            typeof userConfig.server?.cors === 'object' &&
+            userConfig.server.cors.origin
+          )
+        ) {
+          const corsWarnMsgs = [
+            'Detect that CORS options are not set, mf Rsbuild plugin will add default cors header: server.headers["Access-Control-Allow-Headers"] = "*". It is recommended to specify an allowlist of trusted origins in "server.cors" instead.',
+            'View https://module-federation.io/guide/troubleshooting/other.html#cors-warn for more details.',
+          ];
+
+          !isRslib && !isRspress && logger.warn(corsWarnMsgs.join('\n'));
+          config.server.headers['Access-Control-Allow-Origin'] = '*';
+        }
+
+        // For remote modules, Rsbuild should send the ws request to the provider's dev server.
+        // This allows the provider to do HMR when the provider module is loaded in the consumer's page.
+        if (config.server?.port && !config.dev.client?.port) {
+          config.dev.client ||= {};
+          config.dev.client.port = config.server.port;
+        }
+
+        // Change the default assetPrefix to `true` for remote modules.
+        // This ensures that the remote module's assets can be requested by consumer apps with the correct URL.
+        const originalConfig = api.getRsbuildConfig('original');
+        if (
+          originalConfig.dev?.assetPrefix === undefined &&
+          config.dev.assetPrefix === config.server?.base
+        ) {
+          config.dev.assetPrefix = true;
+        }
+      }
+
+      if (isSSR) {
+        if (config.environments?.[SSR_ENV_NAME]) {
+          throw new Error(
+            `'${SSR_ENV_NAME}' environment is already defined. Please use another name.`,
+          );
+        }
+        config.environments![SSR_ENV_NAME] = createSSRREnvConfig(
+          config.environments?.[environment]!,
+          moduleFederationOptions,
+          ssrDir,
+          config,
+          callerName,
+        );
+      }
+    });
+
+    api.modifyEnvironmentConfig((config) => {
+      // Module Federation runtime uses ES6+ syntax,
+      // adding to include and let SWC transform it
+      config.source.include = [
+        ...(config.source.include || []),
+        /@module-federation[\\/]/,
+      ];
+
+      return config;
+    });
+
+    const generateMergedStatsAndManifestOptions: ExposedAPIType = {
+      options: {
+        nodePlugin: undefined,
+        browserPlugin: undefined,
+        distOutputDir: undefined,
+      },
+      isSSRConfig,
+      isRspressSSGConfig,
+    };
+    api.expose(
+      RSBUILD_PLUGIN_MODULE_FEDERATION_NAME,
+      generateMergedStatsAndManifestOptions,
+    );
     api.onBeforeCreateCompiler(({ bundlerConfigs }) => {
       if (!bundlerConfigs) {
         throw new Error('Can not get bundlerConfigs!');
       }
       bundlerConfigs.forEach((bundlerConfig) => {
-        if (!isMFFormat(bundlerConfig)) {
+        if (!isMFFormat(bundlerConfig) && !isRspress) {
           return;
-        } else if (isStoryBook(api.getRsbuildConfig())) {
+        } else if (isStoryBook(originalRsbuildConfig)) {
           bundlerConfig.output!.uniqueName = `${moduleFederationOptions.name}-storybook-host`;
         } else {
           // mf
@@ -98,7 +265,12 @@ export const pluginModuleFederation = (
             moduleFederationOptions,
             bundlerConfig?.optimization?.splitChunks,
           );
+          addDataFetchExposes(
+            moduleFederationOptions.exposes,
+            isSSRConfig(bundlerConfig.name),
+          );
 
+          delete bundlerConfig.optimization?.runtimeChunk;
           const externals = bundlerConfig.externals;
           if (Array.isArray(externals)) {
             const sharedModules = new Set<string>();
@@ -148,8 +320,13 @@ export const pluginModuleFederation = (
             }
           }
 
-          if (!bundlerConfig.output?.chunkLoadingGlobal) {
+          if (
+            !bundlerConfig.output?.chunkLoadingGlobal &&
+            !isSSRConfig(bundlerConfig.name) &&
+            !isRspressSSGConfig(bundlerConfig.name)
+          ) {
             bundlerConfig.output!.chunkLoading = 'jsonp';
+            bundlerConfig.output!.chunkLoadingGlobal = `chunk_${moduleFederationOptions.name}`;
           }
 
           // `uniqueName` is required for react refresh to work
@@ -160,75 +337,72 @@ export const pluginModuleFederation = (
           if (
             !bundlerConfig.plugins!.find((p) => p && p.name === PLUGIN_NAME)
           ) {
+            if (isSSRConfig(bundlerConfig.name)) {
+              generateMergedStatsAndManifestOptions.options.nodePlugin =
+                new ModuleFederationPlugin(
+                  createSSRMFConfig(moduleFederationOptions),
+                );
+              bundlerConfig.plugins!.push(
+                generateMergedStatsAndManifestOptions.options.nodePlugin,
+              );
+              return;
+            } else if (isRspressSSGConfig(bundlerConfig.name)) {
+              const mfConfig = {
+                ...createSSRMFConfig(moduleFederationOptions),
+                // expose in mf-ssg env
+                exposes: {},
+                manifest: false,
+                library: undefined,
+              };
+              patchSSRRspackConfig(
+                bundlerConfig,
+                mfConfig,
+                RSPRESS_SSR_DIR,
+                callerName,
+                false,
+                false,
+              );
+              bundlerConfig.output ||= {};
+              bundlerConfig.output.publicPath = '/';
+              // MF depend on asyncChunks
+              bundlerConfig.output.asyncChunks = undefined;
+              generateMergedStatsAndManifestOptions.options.rspressSSGPlugin =
+                new ModuleFederationPlugin(mfConfig);
+              bundlerConfig.plugins!.push(
+                generateMergedStatsAndManifestOptions.options.rspressSSGPlugin,
+              );
+              return;
+            }
+
+            generateMergedStatsAndManifestOptions.options.browserPlugin =
+              new ModuleFederationPlugin(moduleFederationOptions);
+            generateMergedStatsAndManifestOptions.options.distOutputDir =
+              bundlerConfig.output?.path || '';
             bundlerConfig.plugins!.push(
-              new ModuleFederationPlugin(moduleFederationOptions),
+              generateMergedStatsAndManifestOptions.options.browserPlugin,
             );
           }
         }
       });
     });
 
-    // dev config only works on format: 'mf'
-    api.modifyRsbuildConfig((config) => {
-      // skip storybook
-      if (isStoryBook(config)) {
+    const generateMergedStatsAndManifest = () => {
+      const { nodePlugin, browserPlugin, distOutputDir } =
+        generateMergedStatsAndManifestOptions.options;
+      if (!nodePlugin || !browserPlugin || !distOutputDir) {
         return;
       }
-      // Change some default configs for remote modules
-      if (moduleFederationOptions.exposes) {
-        config.dev ||= {};
-        config.server ||= {};
-        const userConfig = api.getRsbuildConfig('original');
+      updateStatsAndManifest(nodePlugin, browserPlugin, distOutputDir);
+    };
 
-        // Allow remote modules to be loaded by setting CORS headers
-        // This is required for MF to work properly across different origins
-        config.server.headers ||= {};
-        if (
-          !config.server.headers['Access-Control-Allow-Origin'] &&
-          !(
-            typeof userConfig.server?.cors === 'object' &&
-            userConfig.server.cors.origin
-          )
-        ) {
-          const corsWarnMsgs = [
-            'Detect that CORS options are not set, mf Rsbuild plugin will add default cors header: server.headers["Access-Control-Allow-Headers"] = "*". It is recommended to specify an allowlist of trusted origins in "server.cors" instead.',
-            'View https://module-federation.io/guide/troubleshooting/other.html#cors-warn for more details.',
-          ];
-
-          logger.warn(corsWarnMsgs.join('\n'));
-          config.server.headers['Access-Control-Allow-Origin'] = '*';
-        }
-
-        // For remote modules, Rsbuild should send the ws request to the provider's dev server.
-        // This allows the provider to do HMR when the provider module is loaded in the consumer's page.
-        if (config.server?.port && !config.dev.client?.port) {
-          config.dev.client ||= {};
-          config.dev.client.port = config.server.port;
-        }
-
-        // Change the default assetPrefix to `true` for remote modules.
-        // This ensures that the remote module's assets can be requested by consumer apps with the correct URL.
-        const originalConfig = api.getRsbuildConfig('original');
-        if (
-          originalConfig.dev?.assetPrefix === undefined &&
-          config.dev.assetPrefix === config.server?.base
-        ) {
-          config.dev.assetPrefix = true;
-        }
-      }
+    api.onDevCompileDone(() => {
+      generateMergedStatsAndManifest();
     });
 
-    api.modifyEnvironmentConfig((config) => {
-      // Module Federation runtime uses ES6+ syntax,
-      // adding to include and let SWC transform it
-      config.source.include = [
-        ...(config.source.include || []),
-        /@module-federation[\\/]/,
-      ];
-
-      return config;
+    api.onAfterBuild(() => {
+      generateMergedStatsAndManifest();
     });
   },
 });
 
-export { createModuleFederationConfig } from '@module-federation/enhanced';
+export { createModuleFederationConfig } from '@module-federation/sdk';