+++ packages/rsbuild-plugin/src/utils/constant.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_ASSET_PREFIX = '/';