+++ packages/bridge/bridge-react-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @module-federation/bridge-react-webpack-plugin
 
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies [1485fcf]
+  - @module-federation/sdk@0.16.0
+
+## 0.15.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.15.0
+
+## 0.14.3
+
+### Patch Changes
+
+- @module-federation/sdk@0.14.3
+
+## 0.14.2
+
+### Patch Changes
+
+- @module-federation/sdk@0.14.2
+
+## 0.14.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.14.1
+
 ## 0.14.0
 
 ### Patch Changes