+++ packages/runtime-core/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @module-federation/runtime
 
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies [1485fcf]
+  - @module-federation/sdk@0.16.0
+  - @module-federation/error-codes@0.16.0
+
+## 0.15.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.15.0
+- @module-federation/error-codes@0.15.0
+
+## 0.14.3
+
+### Patch Changes
+
+- @module-federation/sdk@0.14.3
+- @module-federation/error-codes@0.14.3
+
+## 0.14.2
+
+### Patch Changes
+
+- @module-federation/sdk@0.14.2
+- @module-federation/error-codes@0.14.2
+
+## 0.14.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.14.1
+- @module-federation/error-codes@0.14.1
+
 ## 0.14.0
 
 ### Minor Changes