+++ packages/runtime-plugins/inject-external-runtime-core-plugin/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @module-federation/inject-external-runtime-core-plugin
 
+## 0.16.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.16.0
+
+## 0.15.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.15.0
+
+## 0.14.3
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.14.3
+
+## 0.14.2
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.14.2
+
+## 0.14.1
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.14.1
+
 ## 0.14.0
 
 ### Patch Changes