+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.16.0
+
+## 0.15.0
+
+## 0.14.3
+
+## 0.14.2
+
+## 0.14.1
+
 ## 0.14.0
 
 ## 0.13.1